+++ src/index.js
@@ -1,5 +1,14 @@
+export { default as classnames } from 'classnames';
+export { default as PropTypes } from 'prop-types';
+export { Fragment, useEffect } from 'react';
+export { render } from 'react-dom';
+export { rendereToString } from 'react-dom/server';
+export { batch, useDispatch, useSelector } from 'react-redux';
+
 export { default as actionCreator } from './action-creator';
+export { default as boundComponent } from './bound-component';
 export { default as concatenateReducers } from './concatenate-reducers';
 export { default as createStore } from './create-store';
+export { default as errorBoundary } from './error-boundary';
 export { default as namespace } from './namespace';
 export { default as namespacedActions } from './namespaced-actions';