+++ RELEASE-NOTES.md
@@ -1,5 +1,11 @@
 # Quoin's React utils library release notes
 
+## 0.0.2 - 2019-12-14
+
+- Adding `boundComponent()`.
+- Exporting `errorBoundary()`.
+- Exporting utility functions from `classnames`, `prop-types`, `react`, `react-dom`, `react-redux`.
+
 ## 0.0.1 - 2019-12-22
 
 - Initial release to test if working.